# core-common-product-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-product-mgmt
  description: Product Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/products
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/products/{productId}:
    get:
      tags:
        - Product
      summary: Get product by ID
      description: Retrieve a specific product by its unique identifier
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved the product
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
    put:
      tags:
        - Product
      summary: Update product
      description: Update the information of an existing product by its unique identifier
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        '200':
          description: Product successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: Invalid product data
        '404':
          description: Product not found
    delete:
      tags:
        - Product
      summary: Delete product
      description: Remove an existing product by its unique identifier
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Product successfully deleted
        '404':
          description: Product not found
...
  /api/v1/products/{productId}/relationships/{relationshipId}:
    get:
      tags:
        - Product Relationship
      summary: Get product relationship by ID
      description: Retrieve a specific relationship record by its unique identifier
      operationId: getRelationshipById
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: relationshipId
          in: path
          description: Unique identifier of the relationship record
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved the product relationship
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductRelationshipDTO'
        '404':
          description: Product relationship not found
    put:
      tags:
        - Product Relationship
      summary: Update product relationship
      description: Update an existing product relationship record
      operationId: updateRelationship
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: relationshipId
          in: path
          description: Unique identifier of the relationship record
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRelationshipDTO'
        required: true
      responses:
        '200':
          description: Product relationship successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductRelationshipDTO'
        '400':
          description: Invalid relationship data
        '404':
          description: Product relationship not found
    delete:
      tags:
        - Product Relationship
      summary: Delete product relationship
      description: Remove an existing product relationship record by its unique identifier
      operationId: deleteRelationship
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: relationshipId
          in: path
          description: Unique identifier of the relationship record
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Product relationship successfully deleted
        '404':
          description: Product relationship not found
...
  /api/v1/products/{productId}/relationships:
    post:
      tags:
        - Product Relationship
      summary: Create product relationship
      description: Create a new relationship record for a specific product
      operationId: createRelationship
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRelationshipDTO'
        required: true
      responses:
        '201':
          description: Product relationship successfully created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductRelationshipDTO'
        '400':
          description: Invalid relationship data
        '404':
          description: Product not found
  /api/v1/products/{productId}/relationships/filter:
    post:
      tags:
        - Product Relationship
      summary: Filter product relationships
      description: Retrieve a paginated list of product relationships based on filtering criteria
      operationId: filterRelationships
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestProductRelationshipDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved filtered relationships
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
        '404':
          description: Product not found
...
  /api/v1/products/filter:
    post:
      tags:
        - Product
      summary: Filter products
      description: Retrieve a paginated list of products based on filtering criteria
      operationId: filterProducts
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestProductDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved filtered products
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
...
components:
  schemas:
    ProductDTO:
      required:
        - tenantId
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        productId:
          type: string
          format: uuid
          readOnly: true
        tenantId:
          type: string
          description: Unique identifier of the tenant instance this product belongs to
          format: uuid
        productCategoryId:
          type: string
          format: uuid
        productType:
          type: string
          enum:
            - FINANCIAL
            - NON_FINANCIAL
        productName:
          type: string
        productCode:
          type: string
        productDescription:
          type: string
        productStatus:
          type: string
          enum:
            - PROPOSED
            - ACTIVE
            - RETIRED
            - DRAFT
        launchDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
...
    ProductRelationshipDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        productRelationshipId:
          type: string
          format: uuid
          readOnly: true
        productId:
          type: string
          format: uuid
        relatedProductId:
          type: string
          format: uuid
        relationshipType:
          type: string
          enum:
            - PRE_REQUISITE
            - COMPLIMENTARY
            - UPGRADE
            - CROSS_SELL
        description:
          type: string
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
    FilterRequestProductRelationshipDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/ProductRelationshipDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for relationships
...
    FilterRequestProductDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/ProductDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for products
...